+++ test/instabug_flutter_test.dart
@@ -253,6 +253,24 @@ void main() {
     ]);
   });
 
+  test(
+      'test setFloatingButtonEdge should be called with arguments floatingButtonEdge and offsetFromTop',
+      () async {
+    const FloatingButtonEdge floatingButtonEdge = FloatingButtonEdge.left;
+    const int offsetFromTop = 300;
+    await BugReporting.setFloatingButtonEdge(floatingButtonEdge, offsetFromTop);
+    final List<dynamic> args = <dynamic>[
+      floatingButtonEdge.toString(),
+      offsetFromTop
+    ];
+    expect(log, <Matcher>[
+      isMethodCall(
+        'setFloatingButtonEdge:withTopOffset:',
+        arguments: args,
+      )
+    ]);
+  });
+
   test('test appendTags should be called with argument List of strings',
       () async {
     const List<String> tags = ['tag1', 'tag2'];